moves more things into src
