+++ src/components/Layout/Layout.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import styles from './Layout.module.css';
 import Link from 'next/link';
 import Header from '../Header';
-import { siteTitle } from '@/constants';
+import { siteTitle } from 'src/constants';
 
 interface LayoutProps {
   children: React.ReactNode;
